Use  PSBTManager on appState environmentObject
This moves part of loading and saving PSBT files away from the SignViewController. In particular it moves opened PSBT to global app state, which paves the way to opening files programatically instead of only through the file picker.

SettingsViewController will crash if you try to access environmentObject(appState), as will any callbacks from it. DispatchQueue() in SignView works around this.