Link to standard Apple EULA instead, as well as MIT
The Apple license seems to mainly apply to what users do on an Apple device, when using the app from the app store. It includes language that carves out open source components. Afaik it does not restrict people from using the source code. So in that sense hopefully the licenses are compatible.